Moved Downgrade classes under Plugins
