+++ src/Config/Rector/Downgrade/Configurators/Plugins/HelloDollyContainerConfigurationService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoP\ExtensionStarter\Config\Rector\Downgrade\Configurators;
+namespace PoP\ExtensionStarter\Config\Rector\Downgrade\Configurators\Plugins;
 
 use PoP\PoP\Config\Rector\Downgrade\Configurators\AbstractExtensionDowngradeContainerConfigurationService;
